+++ CHANGELOG.md
@@ -46,6 +46,7 @@ _None._
 
 ### Internal Changes
 
+- When enabled, `WordPressComRestApi` sends HTTP requests using URLSession instead of Alamofire. [#720]
 - When enabled, `WrodPressOrgXMLRPCApi` sends HTTP requests using URLSession instead of Alamofire. [#719]
 
 ## 13.0.0